requirements.txt
requests
python-telegram-bot
pyyaml